🎨 Refactor types from installers to .shared
