+++ src/installers.shared.ts
@@ -2,19 +2,14 @@ import path from "path";
 import { FileTree, FILETREE_ROOT } from "./filetree";
 import { EXTENSION_NAME_INTERNAL } from "./index.metadata";
 import {
-  LayoutToInstructions,
-  NoInstructions,
-  MaybeInstructions,
   Instructions,
+  LayoutToInstructions,
+  MaybeInstructions,
+  NoInstructions,
 } from "./installers.layouts";
 
 import { VortexApi, VortexInstruction } from "./vortex-wrapper";
-
-export const enum InstallDecision {
-  UserWantsToProceed = "User explicitly wants to proceed with the installation",
-
-  UserWantsToCancel = "User explicitly wants to cancel the installation",
-}
+// Types
 
 // Vortex gives us a 'destination path', which is actually
 // the tempdir in which the archive is expanded into for
